Message Caching added - fast search
Added message caching in message_cache.csv that caches all the messages in the channel where the files are stored, thus dramatically improving search times for older files, as it doesn't have to request every message from discord. Every time the program is run it requests the new messages from discord, it also periodically requests new messages while it's running (every 10 seconds default). The search times are sped up significantly, as the search is done in the file, and it doesn't need to wait around to receive the messages from discord (I realized how inefficient the first version was when I waited 10 minutes for a file download). TL;DR Search is now 10000x or more faster, especially for files that were uploaded at the beginning of the channel.