Fixed issues with encryption
Where do I even start? Encryption was broken in multiple ways. First, only the .part1 file was encrypted, the decryption had issues too. Also, the splitting of the files didn't account for the growth in file size for encrypted files. When the checkbox for encryption is checked it only splits it into 16 mb parts to account for encryption overhead, also all encrypted files are split into parts in the uploads file, and get deleted when the parts are finished uploading. It was the simplest way for me to implement this (or at least the first way I thought about), so if you're uploading a file as encrypted, make sure to have the file space on the drive where the uploaded folder, at least 1.5x your file size,